refactor(backend-service-user): move app modules
