+++ packages/backend-service-user/package.json
@@ -53,7 +53,7 @@
     "format": "prettier --write ./src/**/*.ts",
     "lint": "eslint --ext ts --ignore-path .gitignore ./src",
     "prebuild": "pnpm run build:clean",
-    "serve": "trap '' INT TERM; node lib/app/scripts/bootstrap.js",
+    "serve": "trap '' INT TERM; node lib/app/adapter/nest/scripts/bootstrap.js",
     "test": "jest --config=jest.config.mjs --runInBand",
     "test:integration:js": "pnpm run test:js --selectProjects Integration",
     "test:js": "jest --config=jest.js.config.mjs --runInBand",
